refactor: create utxo directory and move utxo.ts
Ticket: BTC-1092

TICKET: BTC-1092
